test: move cache linker fixtures to dedicated suite
