+++ test/registered/unit/mem_cache/test_unified_cache_linker.py
@@ -1,12 +1,33 @@
+import sys
 import unittest
+from array import array
+from collections import defaultdict
+from dataclasses import replace
 from types import SimpleNamespace
 
 import pytest
 import test_unified_radix_cache_unittest as shared_cache_suite
 import torch
+from test_unified_radix_cache_unittest import (
+    CacheConfig,
+    _device_lock_ref,
+    _device_value,
+    _InsertWalkSuite,
+    _node_children,
+    build_fixture,
+)
 
-from sglang.srt.mem_cache.base_prefix_cache import InsertResult
-from sglang.srt.mem_cache.hicache_storage import PoolName, PoolTransfer
+from sglang.srt.mem_cache.base_prefix_cache import (
+    InitLoadBackParams,
+    InsertResult,
+    MatchPrefixParams,
+)
+from sglang.srt.mem_cache.hicache_storage import (
+    PoolHitPolicy,
+    PoolName,
+    PoolTransfer,
+)
+from sglang.srt.mem_cache.radix_cache import RadixKey
 from sglang.srt.mem_cache.unified_cache.cache_action import (
     ReplaceWriteThroughOnNodeSplit,
 )
@@ -154,6 +175,410 @@ def test_cache_linker_rejects_unsupported_tree_components(component_type):
         UnifiedCacheLinkerWrapper(cache, _FakeLinker())
 
     assert not cache.tree_core.enable_external_cache_linker
+
+
+class _InMemoryUnifiedCacheLinker(UnifiedCacheLinker):
+    """Controllable transport for shared Python/Rust TreeCore tests."""
+
+    def __init__(self, stored_keys=None):
+        self.layer_done_counter = object()
+        self.stored_keys = defaultdict(set) if stored_keys is None else stored_keys
+        self.lookup_calls = []
+        self.offload_calls = []
+        self.pending_offloads = []
+        self.queued_loads = {}
+        self.started_loads = []
+        self.completed_loads = []
+        self.completed_offloads = []
+
+    @staticmethod
+    def _clone_transfer(transfer):
+        return replace(
+            transfer,
+            host_indices=(
+                None if transfer.host_indices is None else transfer.host_indices.clone()
+            ),
+            device_indices=(
+                None
+                if transfer.device_indices is None
+                else transfer.device_indices.clone()
+            ),
+            keys=None if transfer.keys is None else list(transfer.keys),
+            nodes_to_load=(
+                None if transfer.nodes_to_load is None else list(transfer.nodes_to_load)
+            ),
+        )
+
+    @classmethod
+    def _clone_transfers(cls, transfers):
+        return [cls._clone_transfer(transfer) for transfer in transfers]
+
+    def lookup(self, rid, transfers):
+        transfers = self._clone_transfers(transfers)
+        self.lookup_calls.append((rid, transfers))
+        by_pool = {transfer.name: transfer for transfer in transfers}
+        kv = by_pool.get(PoolName.KV)
+        if kv is None or not kv.keys:
+            return []
+
+        restorable = []
+        for prefix_pages in range(1, len(kv.keys) + 1):
+            if not set(kv.keys[:prefix_pages]) <= self.stored_keys[PoolName.KV]:
+                continue
+            for transfer in transfers:
+                if transfer.name == PoolName.KV:
+                    continue
+                if transfer.hit_policy == PoolHitPolicy.TRAILING_PAGES:
+                    window_pages = max(1, len(transfer.keys or ()))
+                    required = kv.keys[
+                        max(0, prefix_pages - window_pages) : prefix_pages
+                    ]
+                else:
+                    required = kv.keys[:prefix_pages]
+                if not set(required) <= self.stored_keys[transfer.name]:
+                    break
+            else:
+                restorable.append(prefix_pages)
+        return restorable
+
+    def load(self, rid, transfers):
+        self.queued_loads[rid] = self._clone_transfers(transfers)
+        return True
+
+    def start_layer_wise_loading(self):
+        if not self.queued_loads:
+            return -1
+        rids = list(self.queued_loads)
+        self.started_loads.append(rids)
+        return len(self.started_loads) - 1
+
+    def cancel_queued_load(self, rid):
+        return self.queued_loads.pop(rid, None) is not None
+
+    def num_completed_loads(self):
+        return len(self.completed_loads)
+
+    def pop_completed_load(self):
+        rids = self.completed_loads.pop(0)
+        for rid in rids:
+            self.queued_loads.pop(rid, None)
+        return rids
+
+    def complete_started_loads(self):
+        self.completed_loads.append(self.started_loads[-1])
+
+    def offload(self, transfers):
+        transfers = self._clone_transfers(transfers)
+        self.offload_calls.append(transfers)
+        self.pending_offloads.append(transfers)
+        return True
+
+    def num_completed_offloads(self):
+        return len(self.completed_offloads)
+
+    def pop_completed_offload(self):
+        return self.completed_offloads.pop(0)
+
+    def complete_next_offload(self, success):
+        transfers = self.pending_offloads.pop(0)
+        if success:
+            for transfer in transfers:
+                self.stored_keys[transfer.name].update(transfer.keys or ())
+        self.completed_offloads.append(success)
+
+    def reset(self):
+        self.queued_loads.clear()
+        self.pending_offloads.clear()
+        self.completed_loads.clear()
+        self.completed_offloads.clear()
+
+    def close(self):
+        self.reset()
+
+
+@unittest.skipUnless(torch.cuda.is_available(), "cache fixtures need CUDA")
+class TestUnifiedCacheLinkerBackends(_InsertWalkSuite):
+    def test_full_offload_load_round_trip_and_dedup(self):
+        cfg = CacheConfig(page_size=2, kv_size=64, max_context_len=64)
+        self.cfg = cfg
+        stored_keys = defaultdict(set)
+
+        producer, producer_allocator, producer_req_pool = build_fixture(cfg)
+        producer_linker = _InMemoryUnifiedCacheLinker(stored_keys)
+        producer.init_cache_linker(producer_linker)
+        tokens = list(range(1, 9))
+
+        inserted = self._insert(producer, producer_allocator, producer_req_pool, tokens)
+        self.assertEqual(len(producer_linker.offload_calls), 1)
+        (kv_offload,) = producer_linker.offload_calls[0]
+        self.assertEqual(kv_offload.name, PoolName.KV)
+        self.assertEqual(
+            kv_offload.keys,
+            producer.tree_core.get_hash_values(inserted.last_device_node),
+        )
+        self.assertTrue(
+            torch.equal(
+                kv_offload.device_indices,
+                _device_value(producer, inserted.last_device_node, ComponentType.FULL),
+            )
+        )
+        self.assertEqual(
+            _device_lock_ref(producer, inserted.last_device_node, ComponentType.FULL),
+            1,
+        )
+        self.assertFalse(
+            producer.tree_core.is_external_cache_stored(inserted.last_device_node)
+        )
+        self.assertEqual(
+            producer.tree_core.get_write_through_pending_id(inserted.last_device_node),
+            inserted.last_device_node,
+        )
+
+        self._insert(producer, producer_allocator, producer_req_pool, tokens)
+        self.assertEqual(len(producer_linker.offload_calls), 1)
+
+        producer_linker.complete_next_offload(True)
+        producer.check_hicache_events()
+        self.assertEqual(
+            _device_lock_ref(producer, inserted.last_device_node, ComponentType.FULL),
+            0,
+        )
+        self.assertTrue(
+            producer.tree_core.is_external_cache_stored(inserted.last_device_node)
+        )
+        self._insert(producer, producer_allocator, producer_req_pool, tokens)
+        self.assertEqual(len(producer_linker.offload_calls), 1)
+
+        consumer, _, consumer_req_pool = build_fixture(cfg)
+        consumer_linker = _InMemoryUnifiedCacheLinker(stored_keys)
+        consumer.init_cache_linker(consumer_linker)
+        req = self._make_req(consumer_req_pool)
+        match = consumer.match_prefix(
+            MatchPrefixParams(key=RadixKey(array("q", tokens)), req=req)
+        )
+        self.assertEqual(match.device_indices.numel(), 0)
+        self.assertEqual(match.host_hit_length, len(tokens))
+        self._apply_match_to_req(req, match)
+
+        loaded, loaded_node = consumer.init_load_back(
+            InitLoadBackParams(
+                best_match_node=match.best_match_node,
+                host_hit_length=match.host_hit_length,
+                req=req,
+            )
+        )
+        self.assertEqual(loaded.numel(), len(tokens))
+        self.assertNotEqual(loaded_node, consumer.root_node_handle())
+        (kv_load,) = consumer_linker.queued_loads[req.rid]
+        self.assertEqual(kv_load.name, PoolName.KV)
+        self.assertEqual(kv_load.keys, kv_offload.keys)
+        self.assertEqual(_device_lock_ref(consumer, loaded_node, ComponentType.FULL), 1)
+
+        self.assertGreaterEqual(consumer.ready_to_load_host_cache(), 0)
+        consumer_linker.complete_started_loads()
+        consumer.check_hicache_events()
+        self.assertEqual(_device_lock_ref(consumer, loaded_node, ComponentType.FULL), 0)
+        final_match = consumer.match_prefix(
+            MatchPrefixParams(key=RadixKey(array("q", tokens)))
+        )
+        self.assertEqual(final_match.device_indices.numel(), len(tokens))
+        self.assertEqual(consumer_linker.offload_calls, [])
+        consumer.sanity_check()
+
+    def test_eagle_lookup_uses_bigram_tail_hashes(self):
+        cfg = CacheConfig(
+            page_size=2,
+            is_eagle=True,
+            kv_size=64,
+            max_context_len=64,
+        )
+        self.cfg = cfg
+        stored_keys = defaultdict(set)
+        tokens = list(range(1, 10))
+
+        producer, producer_allocator, producer_req_pool = build_fixture(cfg)
+        producer_linker = _InMemoryUnifiedCacheLinker(stored_keys)
+        producer.init_cache_linker(producer_linker)
+        self._insert(producer, producer_allocator, producer_req_pool, tokens)
+        self.assertEqual(len(producer_linker.offload_calls), 1)
+        (kv_offload,) = producer_linker.offload_calls[0]
+        self.assertEqual(len(kv_offload.keys), 4)
+        producer_linker.complete_next_offload(True)
+        producer.check_hicache_events()
+
+        consumer, consumer_allocator, consumer_req_pool = build_fixture(cfg)
+        consumer_linker = _InMemoryUnifiedCacheLinker(stored_keys)
+        consumer.init_cache_linker(consumer_linker)
+        consumer.write_through_threshold = sys.maxsize
+        self._insert(
+            consumer,
+            consumer_allocator,
+            consumer_req_pool,
+            tokens[:5],
+        )
+        req = self._make_req(consumer_req_pool)
+        lookup_key = RadixKey(array("q", tokens))
+
+        match = consumer.match_prefix(MatchPrefixParams(key=lookup_key, req=req))
+
+        self.assertTrue(lookup_key.is_bigram)
+        self.assertEqual(match.device_indices.numel(), 4)
+        self.assertEqual(match.host_hit_length, 4)
+        self.assertEqual(len(consumer_linker.lookup_calls), 1)
+        _, transfers = consumer_linker.lookup_calls[0]
+        (kv_lookup,) = transfers
+        self.assertEqual(kv_lookup.keys, kv_offload.keys[2:])
+
+    def test_failed_split_offload_retries_and_reset_releases_locks(self):
+        cfg = CacheConfig(page_size=1, kv_size=64, max_context_len=64)
+        self.cfg = cfg
+        cache, allocator, req_to_token_pool = build_fixture(cfg)
+        linker = _InMemoryUnifiedCacheLinker()
+        cache.init_cache_linker(linker)
+        tokens = [1, 2, 3, 4]
+
+        inserted = self._insert(cache, allocator, req_to_token_pool, tokens)
+        original_node = inserted.last_device_node
+        self.assertEqual(len(linker.offload_calls), 1)
+        original_keys = linker.offload_calls[0][0].keys
+
+        self._insert(cache, allocator, req_to_token_pool, tokens[:2])
+        (parent,) = _node_children(cache, cache.root_node_handle())
+        (child,) = _node_children(cache, parent)
+        self.assertEqual(child, original_node)
+        self.assertEqual(
+            cache.linker.pending_offloads[0].publish_node_ids, [parent, child]
+        )
+        self.assertEqual(cache.tree_core.get_write_through_pending_id(parent), child)
+        self.assertEqual(cache.tree_core.get_write_through_pending_id(child), child)
+        self.assertFalse(cache.tree_core.is_external_cache_stored(parent))
+        self.assertFalse(cache.tree_core.is_external_cache_stored(child))
+
+        linker.complete_next_offload(False)
+        cache.check_hicache_events()
+        for node_id in (parent, child):
+            self.assertIsNone(cache.tree_core.get_write_through_pending_id(node_id))
+            self.assertFalse(cache.tree_core.is_external_cache_stored(node_id))
+            self.assertEqual(_device_lock_ref(cache, node_id, ComponentType.FULL), 0)
+
+        self._insert(cache, allocator, req_to_token_pool, tokens)
+        retry_calls = linker.offload_calls[1:]
+        self.assertEqual(len(retry_calls), 2)
+        self.assertEqual(
+            [key for transfers in retry_calls for key in transfers[0].keys],
+            original_keys,
+        )
+        for _ in retry_calls:
+            linker.complete_next_offload(True)
+        cache.check_hicache_events()
+        for node_id in (parent, child):
+            self.assertIsNone(cache.tree_core.get_write_through_pending_id(node_id))
+            self.assertTrue(cache.tree_core.is_external_cache_stored(node_id))
+            self.assertEqual(_device_lock_ref(cache, node_id, ComponentType.FULL), 0)
+
+        self._insert(cache, allocator, req_to_token_pool, tokens)
+        self.assertEqual(len(linker.offload_calls), 3)
+
+        extended = self._insert(cache, allocator, req_to_token_pool, tokens + [5, 6])
+        pending_node = extended.last_device_node
+        self.assertEqual(len(cache.linker.pending_offloads), 1)
+        self.assertEqual(
+            cache.tree_core.get_write_through_pending_id(pending_node), pending_node
+        )
+        self.assertEqual(_device_lock_ref(cache, pending_node, ComponentType.FULL), 1)
+
+        cache.linker.reset()
+        self.assertEqual(cache.linker.pending_offloads, [])
+        self.assertIsNone(cache.tree_core.get_write_through_pending_id(pending_node))
+        self.assertEqual(_device_lock_ref(cache, pending_node, ComponentType.FULL), 0)
+        cache.sanity_check()
+        cache.reset()
+        cache.sanity_check()
+
+    def test_swa_partial_hit_loads_only_pages_not_adopted_locally(self):
+        cfg = CacheConfig(
+            page_size=1,
+            components=(ComponentType.FULL, ComponentType.SWA),
+            sliding_window_size=2,
+            kv_size=64,
+            max_context_len=64,
+        )
+        self.cfg = cfg
+        stored_keys = defaultdict(set)
+        tokens = list(range(1, 7))
+
+        producer, producer_allocator, producer_req_pool = build_fixture(cfg)
+        producer_linker = _InMemoryUnifiedCacheLinker(stored_keys)
+        producer.init_cache_linker(producer_linker)
+        self._insert(producer, producer_allocator, producer_req_pool, tokens)
+        self.assertGreaterEqual(len(producer_linker.offload_calls), 1)
+        all_keys = []
+        for transfers in producer_linker.offload_calls:
+            offloads = {transfer.name: transfer for transfer in transfers}
+            self.assertEqual(set(offloads), {PoolName.KV, PoolName.SWA})
+            self.assertEqual(offloads[PoolName.KV].keys, offloads[PoolName.SWA].keys)
+            all_keys.extend(offloads[PoolName.KV].keys)
+            producer_linker.complete_next_offload(True)
+        producer.check_hicache_events()
+
+        stored_keys[PoolName.KV].difference_update(all_keys[-2:])
+
+        consumer, consumer_allocator, consumer_req_pool = build_fixture(cfg)
+        consumer_linker = _InMemoryUnifiedCacheLinker(stored_keys)
+        consumer.init_cache_linker(consumer_linker)
+        consumer.write_through_threshold = sys.maxsize
+        self._insert(consumer, consumer_allocator, consumer_req_pool, tokens[:2])
+
+        req = self._make_req(consumer_req_pool)
+        match = consumer.match_prefix(
+            MatchPrefixParams(key=RadixKey(array("q", tokens)), req=req)
+        )
+        self.assertEqual(match.device_indices.numel(), 2)
+        self.assertEqual(match.host_hit_length, 2)
+        self.assertEqual(match.swa_host_hit_length, 2)
+        self._apply_match_to_req(req, match)
+
+        self._insert(consumer, consumer_allocator, consumer_req_pool, tokens[:3])
+        raced_match = consumer.match_prefix(
+            MatchPrefixParams(key=RadixKey(array("q", tokens[:3])))
+        )
+        raced_full = raced_match.device_indices[-1:].clone()
+        raced_swa = consumer_allocator.translate_loc_from_full_to_swa(raced_full)
+
+        loaded, loaded_node = consumer.init_load_back(
+            InitLoadBackParams(
+                best_match_node=match.best_match_node,
+                host_hit_length=match.host_hit_length,
+                req=req,
+            )
+        )
+        self.assertEqual(loaded.numel(), 2)
+        self.assertTrue(torch.equal(loaded[:1], raced_full))
+        load_by_pool = {
+            transfer.name: transfer
+            for transfer in consumer_linker.queued_loads[req.rid]
+        }
+        self.assertEqual(set(load_by_pool), {PoolName.KV, PoolName.SWA})
+        expected_key = all_keys[3]
+        for transfer in load_by_pool.values():
+            self.assertEqual(transfer.keys, [expected_key])
+            self.assertEqual(transfer.device_indices.numel(), 1)
+
+        translated = consumer_allocator.translate_loc_from_full_to_swa(loaded)
+        self.assertTrue(torch.equal(translated[:1], raced_swa))
+        self.assertTrue(
+            torch.equal(translated[-1:], load_by_pool[PoolName.SWA].device_indices)
+        )
+
+        self.assertGreaterEqual(consumer.ready_to_load_host_cache(), 0)
+        consumer_linker.complete_started_loads()
+        consumer.check_hicache_events()
+        final_match = consumer.match_prefix(
+            MatchPrefixParams(key=RadixKey(array("q", tokens[:4])))
+        )
+        self.assertEqual(final_match.device_indices.numel(), 4)
+        self.assertEqual(_device_lock_ref(consumer, loaded_node, ComponentType.FULL), 0)
+        consumer.sanity_check()
 
 
 @unittest.skipUnless(torch.cuda.is_available(), "cache fixtures need CUDA")
